requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai>=1.40.0
google-cloud-firestore==2.16.0
requests==2.32.3
pydantic==2.9.2

